// src/generated/prisma/models/VariantTemplate.ts
export type VariantTemplateWhereInput = {
  AND?: Prisma.VariantTemplateWhereInput | Prisma.VariantTemplateWhereInput[]
  OR?: Prisma.VariantTemplateWhereInput[]
  NOT?: Prisma.VariantTemplateWhereInput | Prisma.VariantTemplateWhereInput[]
  id?: Prisma.StringFilter<"VariantTemplate"> | string
  name?: Prisma.StringFilter<"VariantTemplate"> | string
  description?: Prisma.StringNullableFilter<"VariantTemplate"> | string | null
  createdAt?: Prisma.DateTimeFilter<"VariantTemplate"> | Date | string
  updatedAt?: Prisma.DateTimeFilter<"VariantTemplate"> | Date | string
  options?: Prisma.TemplateOptionListRelationFilter
}